+++ src/peer/negotiator.js
@@ -29,6 +29,7 @@ class Negotiator extends EventEmitter {
    */
   constructor() {
     super();
+    this._offerQueue = [];
     this._isExpectingAnswer = false;
     this._replaceStreamCalled = false;
     this._isNegotiationAllowed = true;
@@ -139,6 +140,13 @@ class Negotiator extends EventEmitter {
     }
 
     this._lastOffer = offerSdp;
+
+    // Enqueue and skip while signalingState is wrong state.
+    // (when room is SFU and there are multiple conns in a same time, it happens)
+    if (this._pc.signalingState === 'have-remote-offer') {
+      this._offerQueue.push(offerSdp);
+      return;
+    }
 
     this._setRemoteDescription(offerSdp)
       .then(() => {
@@ -322,6 +330,18 @@ class Negotiator extends EventEmitter {
 
     pc.onsignalingstatechange = () => {
       logger.log(`signalingState is ${pc.signalingState}`);
+
+      // After signaling state is getting back to 'stable',
+      // apply pended remote offer, which was stored when simultaneous multiple conns happened in SFU room,
+      // Note that this code very rarely applies the old remote offer.
+      // E.g. "Offer A -> Offer B" should be the right order but for some reason like NW unstablity,
+      //      offerQueue might keep "Offer B" first and handle "Offer A" later.
+      if (this._pc.signalingState === 'stable') {
+        const offer = this._offerQueue.shift();
+        if (offer) {
+          this.handleOffer(offer);
+        }
+      }
     };
   }
 
@@ -340,9 +360,9 @@ class Negotiator extends EventEmitter {
     } else {
       if (this._isAddTransceiverAvailable) {
         this._recvonlyState.audio &&
-          this._pc.addTransceiver('audio').setDirection('recvonly');
+          this._pc.addTransceiver('audio', { direction: 'recvonly' });
         this._recvonlyState.video &&
-          this._pc.addTransceiver('video').setDirection('recvonly');
+          this._pc.addTransceiver('video', { direction: 'recvonly' });
         createOfferPromise = this._pc.createOffer();
       } else {
         const offerOptions = {};